WIP: Import of MRP prototype
This is the MRP prototype. It is not usable for anything, it just shows
how the pairing process is performed and how to encrypt/decrypt
messages correctly. It's also a mess and I will continue refactoring
as I get time to do so but I would like to share the code until then.

To try it out, just use:

   atvremote --debug --protocol mrp -a pair

Add --address in case you want to target a specific device (but you
must still use -a to discover the correct port).