Implement more protocol logic in MRP
An abstraction layer, MrpProtocol, has been added that wraps an
MrpConnection. It deals with sending and receiving as well as other
functions (like enabling encryption).

This commit heavily breaks API as "login id" will from now on be called
"device credentials" in the external API. Mostly to have the same
terminology between protocols.